// web/src/api/generated/models/DailyCheckInInsightDayView.ts
/* tslint:disable */
/* eslint-disable */
/**
 * eimir.
 * Application Core. OpenAPI is the authoritative contract.
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { DailyVibe } from './DailyVibe';
import {
    DailyVibeFromJSON,
    DailyVibeFromJSONTyped,
    DailyVibeToJSON,
    DailyVibeToJSONTyped,
} from './DailyVibe';
import type { PartnerVibeProjection } from './PartnerVibeProjection';
import {
    PartnerVibeProjectionFromJSON,
    PartnerVibeProjectionFromJSONTyped,
    PartnerVibeProjectionToJSON,
    PartnerVibeProjectionToJSONTyped,
} from './PartnerVibeProjection';
import type { PartnerEnergyProjection } from './PartnerEnergyProjection';
import {
    PartnerEnergyProjectionFromJSON,
    PartnerEnergyProjectionFromJSONTyped,
    PartnerEnergyProjectionToJSON,
    PartnerEnergyProjectionToJSONTyped,
} from './PartnerEnergyProjection';

/**
 * 
 * @export
 * @interface DailyCheckInInsightDayView
 */
export interface DailyCheckInInsightDayView {
    /**
     * 
     * @type {Date}
     * @memberof DailyCheckInInsightDayView
     */
    checkedOn: Date;
    /**
     * 
     * @type {number}
     * @memberof DailyCheckInInsightDayView
     */
    ownEnergy?: number | null;
    /**
     * 
     * @type {DailyVibe}
     * @memberof DailyCheckInInsightDayView
     */
    ownVibe?: DailyVibe | null;
    /**
     * 
     * @type {PartnerEnergyProjection}
     * @memberof DailyCheckInInsightDayView
     */
    partnerEnergy?: PartnerEnergyProjection | null;
    /**
     * 
     * @type {PartnerVibeProjection}
     * @memberof DailyCheckInInsightDayView
     */
    partnerVibe?: PartnerVibeProjection | null;
}



/**
 * Check if a given object implements the DailyCheckInInsightDayView interface.
 */
export function instanceOfDailyCheckInInsightDayView(value: object): value is DailyCheckInInsightDayView {
    if (!('checkedOn' in value) || value['checkedOn'] === undefined) return false;
    return true;
}

export function DailyCheckInInsightDayViewFromJSON(json: any): DailyCheckInInsightDayView {
    return DailyCheckInInsightDayViewFromJSONTyped(json, false);
}

export function DailyCheckInInsightDayViewFromJSONTyped(json: any, ignoreDiscriminator: boolean): DailyCheckInInsightDayView {
    if (json == null) {
        return json;
    }
    return {
        
        'checkedOn': (new Date(json['checkedOn'])),
        'ownEnergy': json['ownEnergy'] === undefined ? undefined : json['ownEnergy'] === null ? null : json['ownEnergy'],
        'ownVibe': json['ownVibe'] === undefined ? undefined : json['ownVibe'] === null ? null : DailyVibeFromJSON(json['ownVibe']),
        'partnerEnergy': json['partnerEnergy'] === undefined ? undefined : json['partnerEnergy'] === null ? null : PartnerEnergyProjectionFromJSON(json['partnerEnergy']),
        'partnerVibe': json['partnerVibe'] === undefined ? undefined : json['partnerVibe'] === null ? null : PartnerVibeProjectionFromJSON(json['partnerVibe']),
    };
}

export function DailyCheckInInsightDayViewToJSON(json: any): DailyCheckInInsightDayView {
    return DailyCheckInInsightDayViewToJSONTyped(json, false);
}

export function DailyCheckInInsightDayViewToJSONTyped(value?: DailyCheckInInsightDayView | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'checkedOn': value['checkedOn'].toISOString().substring(0,10),
        'ownEnergy': value['ownEnergy'],
        'ownVibe': DailyVibeToJSON(value['ownVibe']),
        'partnerEnergy': PartnerEnergyProjectionToJSON(value['partnerEnergy']),
        'partnerVibe': PartnerVibeProjectionToJSON(value['partnerVibe']),
    };
}