// web/src/api/generated/models/DailyCheckInInsightsView.ts
export interface DailyCheckInInsightsView {
    /**
     * 
     * @type {string}
     * @memberof DailyCheckInInsightsView
     */
    dailyContextTimezone: string;
    /**
     * 
     * @type {Array<DailyCheckInInsightDayView>}
     * @memberof DailyCheckInInsightsView
     */
    days: Array<DailyCheckInInsightDayView>;
    /**
     * 
     * @type {Date}
     * @memberof DailyCheckInInsightsView
     */
    endDate: Date;
    /**
     * 
     * @type {boolean}
     * @memberof DailyCheckInInsightsView
     */
    energyEnabled: boolean;
    /**
     * 
     * @type {Date}
     * @memberof DailyCheckInInsightsView
     */
    startDate: Date;
    /**
     * 
     * @type {DailyCheckInInsightSummaryView}
     * @memberof DailyCheckInInsightsView
     */
    summary: DailyCheckInInsightSummaryView;
    /**
     * 
     * @type {boolean}
     * @memberof DailyCheckInInsightsView
     */
    vibeEnabled: boolean;
}